cais6/outputs/code/gravwavecritique/requirements.txt
numpy
scipy
matplotlib
PyMC3
arviz
astropy
PyYAML
pytest